backend/requirements.txt
# CropMagix Backend Dependencies
# Production-ready requirements

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# API Clients
httpx==0.26.0
google-generativeai==0.8.3

# Data Validation
pydantic==2.5.3

# Environment & Config
python-dotenv==1.0.0

# Production Server
gunicorn==21.2.0

# Utilities
python-multipart==0.0.6

# Image Processing
Pillow>=10.0.0
